requirements.txt
Flask==3.0.3
gunicorn==23.0.0
Werkzeug==3.0.3
PyPDF2==3.0.1
google.genai==1.7.0
google.api_core==2.24.2